+++ restiny/ui/screens/settings_screen.py
@@ -7,6 +7,7 @@ from textual.containers import Horizontal, Vertical
 from textual.screen import ModalScreen
 from textual.widgets import Button, Label, Select
 
+from restiny.entities import Settings
 from restiny.enums import CustomThemes
 
 if TYPE_CHECKING:
@@ -66,8 +67,9 @@ class SettingsScreen(ModalScreen):
 
     @on(Button.Pressed, '#cancel')
     def _on_cancel(self, message: Button.Pressed) -> None:
-        self.dismiss(result=None)
+        self.dismiss(result=False)
 
     @on(Button.Pressed, '#confirm')
     def _on_confirm(self, message: Button.Pressed) -> None:
-        self.dismiss(result={'theme': self.theme_select.value})
+        self.app.settings_repo.set(Settings(theme=self.theme_select.value))
+        self.dismiss(result=True)