+++ restiny/widgets/__init__.py
@@ -3,7 +3,7 @@ This module contains reusable widgets used in the DataFox interface.
 """
 
 from restiny.widgets.collections_tree import CollectionsTree
-from restiny.widgets.confirm_prompt import ConfirmPrompt, ConfirmPromptResult
+from restiny.widgets.confirm_prompt import ConfirmPrompt
 from restiny.widgets.custom_directory_tree import CustomDirectoryTree
 from restiny.widgets.custom_input import CustomInput
 from restiny.widgets.custom_text_area import CustomTextArea
@@ -26,5 +26,4 @@ __all__ = [
     'CustomInput',
     'CollectionsTree',
     'ConfirmPrompt',
-    'ConfirmPromptResult',
 ]