+++ CHANGELOG.md
@@ -5,23 +5,33 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.7.0] - 2025-11-16
+
+### Added
+
+- Postman collection import.
+
+### Changed
+
+- Editing message headers, parameters, or body no longer activates them automatically; activation is now entirely manual.
+
 ## [0.6.3] - 2025-11-11
 
 ### Fixed
 
-- Compatibility with Python 3.11, 3.12, and 3.13
+- Compatibility with Python 3.11, 3.12, and 3.13.
 
 ## [0.6.2] - 2025-11-10
 
 ### Fixed
 
-- Folder update/delete behavior
+- Folder update/delete behavior.
 
 ## [0.6.1] - 2025-11-10
 
 ### Fixed
 
-- Collections related behaviors
+- Collections related behaviors.
 
 ## [0.6.0] - 2025-11-09